demos folder
Created demos folder and moving TensorFlow demo to it
