--- TensorFlow2TensorRTDemo_IUS2023/Cpp/TensorRT_ONNXMNIST/logger.cpp
@@ -1,41 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 1993-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "logger.h"
-#include "ErrorRecorder.h"
-#include "logging.h"
-using namespace nvinfer1;
-SampleErrorRecorder gRecorder;
-namespace sample
-{
-Logger gLogger{Logger::Severity::kINFO};
-LogStreamConsumer gLogVerbose{LOG_VERBOSE(gLogger)};
-LogStreamConsumer gLogInfo{LOG_INFO(gLogger)};
-LogStreamConsumer gLogWarning{LOG_WARN(gLogger)};
-LogStreamConsumer gLogError{LOG_ERROR(gLogger)};
-LogStreamConsumer gLogFatal{LOG_FATAL(gLogger)};
-
-void setReportableSeverity(Logger::Severity severity)
-{
-    gLogger.setReportableSeverity(severity);
-    gLogVerbose.setReportableSeverity(severity);
-    gLogInfo.setReportableSeverity(severity);
-    gLogWarning.setReportableSeverity(severity);
-    gLogError.setReportableSeverity(severity);
-    gLogFatal.setReportableSeverity(severity);
-}
-} // namespace sample
