+++ demos/TensorFlow2TensorRTDemo/Cpp/TensorRT_ONNXMNIST/TensorRT_ONNX_MNIST.sln
@@ -1,0 +1,21 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.33529.622
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TensorRT_ONNX_MNIST", "TensorRT_ONNX_MNIST.vcxproj", "{15A255A6-CDED-467A-B578-1F64F7B9E05F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{15A255A6-CDED-467A-B578-1F64F7B9E05F}.Release|x64.ActiveCfg = Release|x64
+		{15A255A6-CDED-467A-B578-1F64F7B9E05F}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A56458EC-5EA5-4DC7-9158-77D56AEF8ABF}
+	EndGlobalSection
+EndGlobal
