+++ demos/TensorFlow2TensorRTDemo/Cpp/TensorRT_ONNXMNIST/TensorRT_ONNX_MNIST.vcxproj.user
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup />
+</Project>
