+++ demos/TensorFlow2TensorRTDemo/Cpp/sample_mnist_data/5.pgm
