+++ demos/TensorFlow2TensorRTDemo/Cpp/sample_mnist_data/7.pgm
