+++ demos/TensorFlow2TensorRTDemo/Python/simple_nn/variables/variables.data-00000-of-00001
